import: Move SubscriberHandler to import_util.
